Move JavaScript tests to test/js
