+++ test/js/bwtest.html
@@ -12,4 +12,4 @@
     open browser - file:///.../node-webrtc/test/bwtest.html
 
 -->
-<script src="../build/bwtest-browserify.js"></script>
+<script src="../../build/bwtest-browserify.js"></script>
