+++ test/js/helpers/capture-candidates.js
