+++ test/js/sessiondesc.js
@@ -2,7 +2,7 @@
 
 var test = require('tape');
 
-var wrtc = require('..');
+var wrtc = require('../..');
 
 var RTCPeerConnection     = wrtc.RTCPeerConnection;
 var RTCSessionDescription = wrtc.RTCSessionDescription;
